requirements.txt
asgiref==3.2.7
astroid==2.4.1
autopep8==1.5.2
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.1
dj-database-url==0.5.0
Django==3.0.6
djangorestframework==3.11.0
drf-yasg==1.17.1
gunicorn==20.0.4
idna==2.9
inflection==0.4.0
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.3
psycopg2-binary==2.8.5
pycodestyle==2.6.0
PyJWT==1.7.1
pylint==2.5.2
pyparsing==2.4.7
pytz==2020.1
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9
whitenoise==5.0.1
wrapt==1.12.1
